feat(core): Map native integer metadata onto FoundryScript types
A `PropertyInfo` transports only the Variant carrier, so a native
integer's exact width lived nowhere the analyzer could read it. Argument
metadata is now available in every build that can compile Foundry source,
unsigned C++ integers declare the unsigned carrier alongside their width,
and one metadata-aware conversion helper decodes every native boundary.

- `MethodBind::get_argument_meta()` moves out of `DEBUG_ENABLED` as
  read-only metadata; no debug-only diagnostic storage changes.
- `GetTypeInfo<uintN_t>` declares `Variant::UINT`, so a bound result keeps
  the whole unsigned range instead of being retagged signed. Nominal
  wrappers such as `ObjectID` keep their explicit signed binding.
- Argument validation normalizes either integer carrier in both
  directions, and a default argument is retagged onto its parameter's
  declared carrier when it is representable there.
- `FSAnalyzer::type_from_property()` takes the declaring bind's metadata:
  int8/16/32 map to signed constraints rendered as `int`, uint8/16/32 to
  unsigned ones rendered as `uint`, and a missing width to `long`/`ulong`.
  Method parameters, returns, getters, setters, properties, and signals
  all decode through it.
- An int-backed enum satisfies an unsigned integer slot, since native flag
  parameters are unsigned while the enum naming their values is not.

Constant arguments outside a native parameter's range now fail during
analysis, and a dynamic value crossing carriers needs a checked cast that
fails before the native method is invoked.

Closes #1570

+++ editor/debugger/editor_debugger_inspector.cpp
@@ -144,7 +144,7 @@ void EditorDebuggerInspector::_notification(int p_what) {
 		} break;
 
 		case NOTIFICATION_ENTER_TREE: {
-			variables->remote_object_ids.append(0);
+			variables->remote_object_ids.append(uint64_t(0));
 			edit(variables);
 		} break;
 	}
@@ -167,9 +167,7 @@ EditorDebuggerRemoteObjects *EditorDebuggerInspector::set_objects(const Array &p
 		SceneDebuggerObject obj;
 		obj.deserialize(arr);
 		if (obj.id.is_valid()) {
-			// The element type of a `TypedArray<uint64_t>` is declared `Variant::INT`, so the
-			// instance id is carried nominally rather than in the unsigned carrier.
-			ids.push_back((int64_t)(uint64_t)obj.id);
+			ids.push_back((uint64_t)obj.id);
 			objects.push_back(obj);
 		}
 	}
